pyproject.toml
[project]
name = "ascension-clips"
version = "2.0.0"
description = "AI-powered Twitch clipping automation for streamers"
authors = [{name = "AscensionClips Team"}]
readme = "README_DEPLOYMENT.md"
requires-python = ">=3.10"
dynamic = ["dependencies"]

[project.scripts]
start-api = "scripts.start_api:main"

[tool.poetry]
packages = [{include = "src"}, {include = "scripts"}]

[tool.poetry.dependencies]
python = "^3.10"
# Core dependencies
yt-dlp = ">=2023.12.30"
requests = ">=2.31.0"
twitchio = ">=2.8.0"
streamlink = ">=6.0.0"
python-dotenv = ">=1.0.0"
# API framework
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
pydantic = ">=2.0.0"
PyJWT = ">=2.8.0"
# Database
supabase = ">=2.0.0"
# ML/AI dependencies
numpy = ">=1.24.0"
scikit-learn = ">=1.3.0"
openai = ">=1.0.0"
torch = ">=2.0.0"
faster-whisper = "*"
transformers = "*"
librosa = ">=0.10.0"
soundfile = ">=0.12.0"
joblib = ">=1.3.0"
pyannote-audio = "*"
# Google APIs
google-api-python-client = "==2.100.0"
google-auth-oauthlib = "==1.1.0"
google-auth-httplib2 = "==0.1.1"
requests-oauthlib = "==1.3.1"
# Background tasks
celery = "==5.3.4"
redis = "==5.0.1"
# System utilities
psutil = ">=5.9.0"
# LLM monitoring (optional - package not yet available on PyPI)
# llmobserve = ">=0.1.0"  # Install manually when available: pip install llmobserve

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
